requirements.txt
django
gunicorn
whitenoise
requests
psycopg
urllib3
python-dotenv
asgiref
certifi
charset-normalizer
dj-database-url
sqlparse
pytest
django-webtest
coverage
django-discover-runner
factory_boy
model_mommy
mock
djangorestframework
django-safedelete==1.3.1
psycopg2-binary